aws light restruct (user facing stuff, the contract of the library)
